backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
greenlet==3.0.3
pydantic==2.8.2
pydantic-settings==2.3.4
redis==5.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.1.19
python-dotenv==1.0.1
httpx==0.27.0
structlog==24.2.0
python-multipart==0.0.9
bcrypt==4.0.1
pyotp==2.9.0
